chore: combine source
