+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.9.2",
   "description": "Blazing-fast binary parser builder",
   "main": "dist/binary_parser.js",
-  "module": "dist/esm/binary-parser.mjs",
+  "module": "dist/esm/binary_parser.mjs",
   "devDependencies": {
     "@types/node": "^16.9.1",
     "mocha": "^9.1.1",
@@ -14,13 +14,13 @@
   "exports": {
     ".": {
       "import": "./dist/esm/binary_parser.mjs",
-      "require": "./dist/binary-parser.js"
+      "require": "./dist/binary_parser.js"
     },
     "./*": "./*"
   },
   "scripts": {
     "build": "tsc && npm run build:esm",
-    "build:esm": "tsc --target esnext --module esnext --outDir dist/esm && node rename",
+    "build:esm": "tsc --target esnext --module esnext --outDir dist/esm && mv dist/esm/binary_parser.js dist/esm/binary_parser.mjs",
     "fmt": "prettier --write \"{lib,example,test,benchmark}/**/*.{ts,js}\"",
     "check-fmt": "prettier --list-different \"{lib,example,test,benchmark}/**/*.{ts,js}\"",
     "test": "mocha",
